Action

Documentation

data Request :: *

Information on the request sent by the client. This abstracts away the details of the underlying implementation.

data RequestContext

data Handler a

The monad in which route handlers run. At the top, each route Action returns a Handler Response

data Action

This type captures both the authentication needs and the handler for a route.

It extends ActionRoute, which already has most info about a route and how to serve it, to include the authentication requirement.

data Response :: *

response :: ResponseData r => Status -> ResponseHeaders -> r -> Response

emptyResponse :: Status -> ResponseHeaders -> Response

redirectRouteResponse :: Status -> ResponseHeaders -> RouteAction r a -> r -> Request -> Response

Redirect a request to a new route

otherRouteResponse :: ResponseHeaders -> RouteAction r a -> r -> Request -> Response

Redirect with HTTP code 303

forbiddenResponse :: RequestContext -> Response

HTTP code 403

notFoundResponse :: RequestContext -> Response

HTTP code 404

okResponse :: ResponseData r => ResponseHeaders -> r -> Response

result :: MonadIO m => Response -> m a

Short circuit immediately, returning the given Reponse.

FIXME: Rather than implementing this using exceptions, could we either use a left-biased Alternative or simply use a better procedural style?

maybeAction :: Maybe a -> Handler a

Fail with 404 if not Just

module Action.Route

withAuth

Arguments

:: Handler Response
The handler for the route

-> Action
The bundled action

Tag an Action as needing to know whether or not the user is authenticated. This simply triggers an extra session lookup in actionApp.

withoutAuth :: Handler Response -> Action

Tag an Action as not needing auth. See note at withAuth.

Building the application

newtype WaiRouteApp

Constructors

WaiRouteApp Application

newtype ActionRouteApp

Constructors

ActionRouteApp Application

actionRouteApp

Arguments

:: RouteMap Action
The original route map. Now partially replaced by Wai Routes

-> WaiRouteApp
The newer Wai Route-based Application

-> Service
System capabilities

-> ActionRouteApp
The actual web app

The second level of the Databrary 'web framework'. Makes a (wrapped) Wai Application given a route map, a hatch into the Wai Route fallback, and the already-generated system capabilities.

Most routes are served by web-inv-route—style ActionRoutes, but some (~30%) have been converted to Wai Routes.