Controller.Authorize

Documentation

viewAuthorize :: ActionRoute (API, PartyTarget, AuthorizeTarget)

postAuthorize :: ActionRoute (API, PartyTarget, AuthorizeTarget)

Either create a new authorization request from PartyTarget child to a parent or updatecreatereject with validation errors an authorization request to the PartyTarget parent from a child

deleteAuthorize :: ActionRoute (API, PartyTarget, AuthorizeTarget)

If present, delete either a prior request for authorization. The authorization to delete can be specified from the child perspective (child party is pathPartyTarget) or the parent perspective (parent party is pathPartyTarget). Inform all relevant parties that the authorization has been deleted.

postAuthorizeNotFound :: ActionRoute PartyTarget

During registration and when requesting additional sponsors, if the target parent party doesn't exist in Databrary yet, this route enables a user to submit some information on which target parent (AI or institution) they are seeking, to trigger an email to the Databrary site admins, with the hope that the site admins are able to manually get the intended parties into Databrary.