Controller.Party

Documentation

getParty :: Maybe Permission -> PartyTarget -> Handler Party

Ensure the current user has the requested permission for the given party.

viewParty :: ActionRoute (API, PartyTarget)

viewPartyEdit :: ActionRoute PartyTarget

viewPartyCreateHandler :: Action

viewPartyDelete :: ActionRoute (Id Party)

postParty :: ActionRoute (API, PartyTarget)

createParty :: ActionRoute API

Create a new party, starting from blankParty, then overlaying data extracted from request. Since this overlaps with how registration creates an account/party combo, this is only used when creating parties manually by a superadmin, such as creating institution parties which don't have an account.

deleteParty :: ActionRoute (Id Party)

viewAvatar :: ActionRoute (Id Party)

queryParties :: ActionRoute API

Handle route to find parties by the provided PartyFilter.

adminParties :: ActionRoute ()

adminPartiesHandler :: Action

csvPartiesHandler :: Action

csvDuplicatePartiesHandler :: Action