Controller.Upload

Documentation

uploadStart :: ActionRoute (Id Volume)

uploadChunk :: ActionRoute ()

testChunk :: ActionRoute ()

for testing

createUploadSetSize :: (MonadHas Entropy c m, MonadDB c m, MonadHasIdentity c m, MonadStorage c m) => Volume -> UploadStartRequest -> m Upload

data UploadStartRequest

Constructors

UploadStartRequest ByteString Int64

writeChunk

Arguments

:: Int64
Offset to start writing block into

-> Word64
Length of block to be written

-> RawFilePath
The target file to write into

-> IO ByteString
The data source that provides chunks of data for writing

-> IO Word64
number of bytes written

Write one contiguous block of data to a file