Data.Csv.Contrib

Documentation

getHeaders :: Header -> [ByteString]

extractColumnsInitialRows :: Int -> Header -> Vector NamedRecord -> [(ByteString, [ByteString])]

extractColumnsDistinctSample :: Int -> Header -> Vector NamedRecord -> [(ByteString, [ByteString])]

extractColumnDefaulting :: ByteString -> Vector NamedRecord -> [ByteString]

extractColumn :: ByteString -> Vector NamedRecord -> (Maybe ByteString -> a) -> [a]

decodeCsvByNameWith :: (NamedRecord -> Parser a) -> ByteString -> Either String (Header, Vector a)

parseCsvWithHeader :: ByteString -> Either String (Header, Vector NamedRecord)

removeBomPrefixText :: Text -> Text

some programs introduce a byte order mark when generating a CSV, remove this per cassava issue recipe

removeBomPrefix :: ByteString -> ByteString

some programs introduce a byte order mark when generating a CSV, remove this per cassava issue recipe

repairDuplicateLineEndings :: ByteString -> ByteString

fix duplicate line endings, unclear if SPSS or Excel introduces them

repairCarriageReturnOnly :: ByteString -> ByteString

only fix newlines for bizarre macOS endings that use r instead of rn