Databrary.API

Description

This module exists purely to break a cycle between Databrary.Action and Databrary.Routes. I would like to see its contents folded into Databrary.Routes eventually, if possible.

Documentation

type ServantAPI = PreferAngularAPI :<|> ActionRouteAPI

The Databrary API, as described by Servant. It's called ServantAPI, instead of DatabraryAPI, because it's currently the *third* API description for Databrary in this codebase.

type ActionRouteAPI = Raw

type PreferAngularAPI = Header "User-Agent" Text :> (QueryParam "js" Int :> EmptyAPI)

Routes that fire up the Angular SPA if possible.