Databrary.Action.Run

Documentation

data Action

runAction :: Service -> Action -> Application

Run the requested action with some resolved identity, send response, with some header decorations.

forkAction :: Handler a -> RequestContext -> (Either SomeException a -> IO ()) -> IO ThreadId

withAuth :: Handler Response -> Action

withoutAuth :: Handler Response -> Action

withReAuth :: SiteAuth -> Handler a -> Handler a

This may be like a su that allows running an action as a different SiteAuth.

FIXME: It is a little annoying that Identified carries a session, while ReIdentified carries a SiteAuth.