Databrary.Context

Documentation

data ActionContext

This is the context for when you don't have an identity, but you have a fully initialized, "command line" access to the system.

Constructors

ActionContext

Fields

contextService :: !Service

All initialized services; "the imperative shell"

contextTimestamp :: !Timestamp

When the ContextM action is running (i.e., NOW)

contextResourceState :: !InternalState

Optimization for MonadResource

contextDB :: !DBConn

The specific connection chosen for the running action?

type ContextM a = ReaderT ActionContext IO a

FIXME: New name?

runContextM :: ContextM a -> Service -> IO a

Perform an atomic action without an identity with a guaranteed database connection and a fixed version of now.

newtype BackgroundContext

A ActionContext with no Identity.

Constructors

BackgroundContext

Fields

backgroundContext :: ActionContext

type BackgroundContextM a = ReaderT BackgroundContext IO a

withBackgroundContextM :: BackgroundContextM a -> ContextM a