Databrary.Model.Authorize

Documentation

module Databrary.Model.Authorize.Types

selfAuthorize :: Party -> Authorize

lookupAuthorizedChildren :: (MonadDB c m, MonadHasIdentity c m) => Party -> Maybe Permission -> m [Authorize]

lookupAuthorizedParents :: (MonadDB c m, MonadHasIdentity c m) => Party -> Maybe Permission -> m [Authorize]

lookupAuthorize :: (MonadDB c m, MonadHasIdentity c m) => Party -> Party -> m (Maybe Authorize)

Attempt to find an authorization request or grant from the child party to the granting parent party

lookupAuthorizeParent :: (MonadDB c m, MonadHasIdentity c m) => Party -> Id Party -> m (Maybe Authorize)

lookupAuthorization :: (MonadDB c m, MonadHasIdentity c m) => Party -> Party -> m Authorization

changeAuthorize :: MonadAudit c m => Authorize -> m ()

removeAuthorize :: MonadAudit c m => Authorize -> m Bool

authorizeExpired :: Authorize -> Timestamp -> Bool

authorizeActive :: Authorize -> Timestamp -> Bool

authorizeJSON :: ToObject o => Authorize -> o

lookupAuthorizeActivity :: (MonadDB c m, MonadHasIdentity c m) => Int -> m [(Timestamp, Party)]