Databrary.Model.Format

Documentation

module Databrary.Model.Format.Types

mimeTypeTop :: ByteString -> ByteString

Parse a full mimetype value into just its category or type

mimeTypeSub :: ByteString -> ByteString

Parse a full mimetype value into just its subtype

mimeTypeTopCompare :: ByteString -> ByteString -> Ordering

Establish a relative order between to specific mimetype values ...

allFormats :: [Format]

Harcoded list of all formats recognized by Databrary for uploading TODO: db coherence

getFormat :: Id Format -> Maybe Format

getFormat' :: Id Format -> Format

getFormatByExtension :: ByteString -> Maybe Format

addFormatExtension :: RawFilePath -> Format -> RawFilePath

getFormatByFilename :: RawFilePath -> Maybe Format

dropFormatExtension :: Format -> RawFilePath -> RawFilePath

videoFormat :: Format

Blessed video format, used as output from transcoding, as well as understood by front end for playback

imageFormat :: Format

audioFormat :: Format

Blessed audio format, used as output from transcoding, as well as understood by front end for playback

formatIsImage :: Format -> Bool

Is this an image format

formatTranscodable :: Format -> Maybe Format

If the format can (or should be) transcoded into the blessed internal format, then provide the format it will be transcoded into

formatSample :: Format -> Maybe Format

For formats where we can produce samples, determine the type for sample output

formatJSON :: ToObject o => Format -> Record (Id Format) o

Convert a Format value into the shape of JSON to be produced. Arbitrarily only show the first extension if multiple extensions are present

formatIsAV :: Format -> Bool

Is this an Audio or Video format

formatNotAV :: Format -> Bool