Databrary.Model.GeoNames

Documentation

data GeoName

values retrieved from geonames service. large db of place names including countries

Constructors

GeoName

Fields

geoNameId :: !(Id GeoName)

identifier from geonames service

geoName :: !Text

human readable place name

geoNameUS :: GeoName

hardcoded value for US geoname place

parseGeoNameRef :: String -> Maybe (Id GeoName)

Extract geoname identifier from a geoname place url

lookupGeoName :: Id GeoName -> HTTPClient -> IO (Maybe GeoName)

Perform a geoname Id based place lookup to get the corresponding place name, parsing the response into a GeoName value

parseGeoName :: Value -> Parser GeoName

Parse the json response from a geoname id based place lookup into a GeoName value