Databrary.Model.Party.Types

Documentation

data PartyRow

Constructors

PartyRow

Fields

partyId :: Id Party

partySortName :: Text

partyPreName :: Maybe Text

partyORCID :: Maybe ORCID

partyAffiliation :: Maybe Text

partyURL :: Maybe URI

data Party

Represents users, institutions, labs, *and* groups.

Constructors

Party

Fields

partyRow :: !PartyRow

partyAccount :: Maybe Account

partyPermission :: Permission

permission current user has over this party

partyAccess :: Maybe Access

direct authorization this party has granted to current user

data Account

Constructors

Account

Fields

accountEmail :: ByteString

accountParty :: Party

data SiteAuth

TODO: clarify. This is not necessarily a session, but... some user (human being) who has been granted access to the site. There is a corner case indirection because sometimes a job runs a human being.

Constructors

SiteAuth

Fields

siteAccount :: Account

maybe should be Party (for nobody)

accountPasswd :: Maybe ByteString

siteAccess :: Access

Still figuring out what an Access is.

nobodySiteAuth :: SiteAuth

blankParty :: Party

Uninitialized Party object to be used in creating new parties (and accounts)

blankAccount :: Account

Uninitialized Account object to be used in creating new accounts