Databrary.Model.Permission

Documentation

module Databrary.Model.Permission.Types

permissionVIEW :: Permission

Level at which things become visible.

permissionPRIVATE :: Permission

Alias for READ. Grants full access to private data, bypassing consent permissions.

readPermission :: Release -> Permission

The necessary permission level to read a data object with the given release. Equivalent to the SQL function read_permission.

readRelease :: Permission -> Maybe Release

The most restrictive data release level that the current user may access under the given permission. Equivalent to the SQL function read_release. Inverse of readPermission module meaning of Nothing.

dataPermission :: (Has Release a, Has Permission a) => a -> Permission

accessJSON :: ToObject o => Access -> o