Databrary.Ops

Description

FIXME: There is a lot of duplication of standard library tools in here.

Documentation

(<?) :: Alternative f => a -> Bool -> f a infixl 1

flip '(?>)'

(?>) :: Alternative f => Bool -> a -> f a infixr 1

'($>)' . guard

(?!>) :: Alternative f => Bool -> a -> f a infixr 1

'(?>)' . not

(?$>) :: (Applicative m, Alternative f) => Bool -> m a -> m (f a) infixr 1

liftM . (?>)

(?!$>) :: (Applicative m, Alternative f) => Bool -> m a -> m (f a) infixr 1

'(?$>)' . not

rightJust :: Either a b -> Maybe b

fromMaybeM :: Monad m => m a -> Maybe a -> m a

orElseM :: Monad m => Maybe a -> m (Maybe a) -> m (Maybe a) infixl 3

flatMapM :: Monad m => (a -> m (Maybe b)) -> Maybe a -> m (Maybe b)

groupTuplesBy :: (a -> a -> Bool) -> [(a, b)] -> [(a, [b])]

mergeBy :: (a -> a -> Ordering) -> [a] -> [a] -> [a]

Merge two ordered lists using the given predicate, removing EQ "duplicates" (left-biased)