Model.Access

Description

A collection of access request methods.

Documentation

accessSlot :: (MonadDB c m, MonadHasIdentity c m) => Permission -> Id Slot -> m (AccessResult Slot)

Lookup a Slot by its Id, requesting the given permission.

accessVolume :: (MonadDB c m, MonadHasIdentity c m) => Permission -> Id Volume -> m (AccessResult Volume)

Lookup a Volume by its Id, requesting the given permission.

data AccessResult a

Captures possible request responses. NOTE: This was designed to mimic existing code and responses. LookupFailed does NOT mean "does not exist". It means that lookupVolume (for example) returned Nothing. This could mean either the id is a valid id, or the user doesn't have access to the volume.

TODO: Monad Transformer?

Constructors

LookupFailed

AccessDenied

AccessResult a