Model.Permission.Types

Documentation

data Permission

Constructors

PermissionNONE

PermissionPUBLIC

PermissionSHARED

PermissionREAD

PermissionEDIT

PermissionADMIN

data Access

TODO: Figure out what this thing represents

Constructors

Access

Fields

accessSite' :: !Permission

A given acting user's permission level on on the databrary site's data (e.g. volumes). The permission is computed through considering inherited permissions from the databrary site (party 0) down to the acting user.

accessMember' :: !Permission

??

accessPermission' :: Access -> Permission

accessSite :: Has Access a => a -> Permission

accessMember :: Has Access a => a -> Permission

accessPermission :: Has Access a => a -> Permission

data PublicPolicy

A PublicPolicy represents a set of rules that customize the public viewer role for a given volume. Restricted is the only current policy. It signifies hiding all data, except high level summary information and highlights. The word policy is a reference to the term used in attribute-based access control.

Constructors

PublicRestrictedPolicy

PublicNoPolicy

data SharedPolicy

A SharedPolicy is the same as PublicPolicy currently, but applied to the shared viewer role.

Constructors

SharedRestrictedPolicy

SharedNoPolicy

data VolumeRolePolicy

A user's effective access to a given volume.

Constructors

RoleNone

RolePublicViewer PublicPolicy

RoleSharedViewer SharedPolicy

RoleReader

RoleEditor

RoleAdmin

extractPermissionIgnorePolicy :: VolumeRolePolicy -> Permission

Transition function used until all call sites take into Policy value into consideration.

Orphan instances

PGType "permission"

Methods

pgTypeName :: PGTypeName "permission" -> String

pgBinaryColumn :: PGTypeEnv -> PGTypeName "permission" -> Bool