Model.Record.Types

Documentation

data RecordRow

Constructors

RecordRow

Fields

recordId :: Id Record

recordCategory :: Category

data Record

Constructors

Record

Fields

recordRow :: !RecordRow

recordMeasures :: Measures

recordRelease :: Maybe Release

recordVolume :: Volume

data ParticipantRecord

Constructors

ParticipantRecord

Fields

prdId :: FieldUse ByteString

prdInfo :: FieldUse ByteString

prdDescription :: FieldUse ByteString

prdBirthdate :: FieldUse Day

prdGender :: FieldUse ByteString

prdRace :: FieldUse ByteString

prdEthnicity :: FieldUse ByteString

prdGestationalAge :: FieldUse Double

prdPregnancyTerm :: FieldUse ByteString

prdBirthWeight :: FieldUse Double

prdDisability :: FieldUse ByteString

prdLanguage :: FieldUse ByteString

prdCountry :: FieldUse ByteString

prdState :: FieldUse ByteString

prdSetting :: FieldUse ByteString

data FieldUse a

States for fields within ParticipantRecord.

We keep both the parsed and raw values for data for now...

Constructors

FieldUnused
Unused/not supplied

FieldEmpty
Supplied, but empty

Field MeasureDatum a
The raw value and its converted form deriving (Show, Eq, Ord, Functor)

getRecordVolumePermission :: Record -> VolumeRolePolicy

data Measure

Constructors

Measure

Fields

measureRecord :: Record

measureMetric :: Metric

measureDatum :: !MeasureDatum

type Measures = [Measure]

blankRecord :: Category -> Volume -> Record