Model.Slot

Documentation

module Model.Slot.Types

lookupSlot :: (MonadDB c m, MonadHasIdentity c m) => Id Slot -> m (Maybe Slot)

Look up a Slot by its Id, gated by the running Identity's permission to view the Slot's Container's Volume. :)

lookupContainerSlot :: (MonadDB c m, MonadHasIdentity c m) => Id Container -> m (Maybe Slot)

Look up a Slot by its Container's Id, gated by the running Identity's permission to view the Volume containing the Container (which contains the Slot).

auditSlotDownload :: MonadAudit c m => Bool -> Slot -> m ()

slotJSON :: ToObject o => Slot -> Record (Id Container) o