Model.Volume

Documentation

module Model.Volume.Types

coreVolume :: Volume

lookupVolume :: (MonadDB c m, MonadHasIdentity c m) => Id Volume -> m (Maybe Volume)

getVolume2

Arguments

:: (MonadDB c ctrl, MonadHasIdentity c ctrl)

=> Permission
Desired permission level

-> Id Volume
Id of Volume to get

-> ctrl (LookupResult Volume)

Get a Volume, if it is available to the given Identity with the desired Permission.

data LookupResult a

Captures possible database responses. NOTE: This was designed to mimic existing code and responses. LookupFailed does NOT mean "does not exist". It means that lookupVolume (for example) returned Nothing. This could mean either the id is a valid id, or the user doesn't have access to the volume.

Constructors

LookupFailed

LookupDenied

LookupFound a

changeVolume :: MonadAudit c m => Volume -> m ()

addVolume :: MonadAudit c m => Volume -> m Volume

auditVolumeDownload :: MonadAudit c m => Bool -> Volume -> m ()

data VolumeFilter

Constructors

VolumeFilter

Fields

volumeFilterQuery :: Maybe String

volumeFilterParty :: Maybe (Id Party)

volumeFilterPaginate :: Paginate

findVolumes :: (MonadHasIdentity c m, MonadDB c m) => VolumeFilter -> m [Volume]

getVolumeAlias :: Volume -> Maybe Text

volumeRowJSON :: ToObject o => VolumeRow -> Record (Id Volume) o

volumeJSON :: ToObject o => Volume -> Maybe [VolumeAccess] -> Record (Id Volume) o

volumeJSONSimple :: ToObject o => Volume -> Record (Id Volume) o

updateVolumeIndex :: MonadDB c m => m ()