Routes

Description

This module describes the routes served by Databrary.

It is a tale of systems that evolve over time. Writing these words on 2018-05-23, I am beginning to serve routes with Servant. Meanwhile, ~90 routes are still served by the original system, web-inv-routes; and ~30 are served by Wai.Route as a temporary stopgap.

This module glues the API description to a particular service implementation. See API for a pure description of the Servant-described API.

Temporary measures: Wai.Route

routeMapWai :: Service -> [(ByteString, Handler IO)]

Map of route actions managed by Wai Routes.

OG route descriptions: web-inv-routes

routeMapInvertible :: RouteMap Action

Map of route actions handled by web-inv-routes.