Service.Mail

Documentation

initMailer :: Mailer

Make an instance of Mailer with default sendMail implementation, for use in settings where mailer is not being mocked

data Mailer

Parts of mailer that can be mocked

Constructors

Mailer

Fields

mlr :: MailHost -> MailPort -> MailUser -> MailPass -> Mail -> IO ()

type MonadMail c m = (MonadLog c m, MonadHas Mailer c m)

sendMail :: MonadMail c m => [Either ByteString Account] -> [Either ByteString Account] -> Text -> Text -> m ()

wrapText :: Int64 -> Text -> Text

Wrap text to the given line length where possibleby changing some ' ' to '\n'. >>> wrapText 10 $ TL.pack "hello there this is a test wherethereareexactlyvery long.nLines.nnThing with multiple.n" "hellonthere thisnis a testnwherethereareexactlyverynlong.nLines.nnThing withnmultiple.n"