Control.Invert

Documentation

data InvertM b a

runInvert :: InvertM b () -> IO (IO (Maybe b))

Convert an InvertM action into an IO action that, when called repeatedly, returns Just for each give, and Nothing after the action completes.

give :: b -> InvertM b ()