Databrary.Action.Route

Documentation

data Method :: *

Standard HTTP methods. These are defined a number of places already, but the http-types version (which is the only thing we import by default) is too cumbersome.

Constructors

OPTIONS

GET

HEAD

POST

PUT

DELETE

TRACE

CONNECT

PATCH

ExtensionMethod ~ByteString

type ActionRoute a = RouteAction a Action

actionURL :: Maybe Request -> RouteAction r a -> r -> Query -> Builder

actionURI :: Maybe Request -> RouteAction r a -> r -> Query -> URI

actionMethod :: RouteAction r a -> r -> Method

action :: Method -> PathParser r -> (r -> a) -> RouteAction r a

multipartAction :: RouteAction q a -> RouteAction q a

data API

Constructors

HTML

JSON

pathHTML :: PathParser ()

pathJSON :: PathParser ()

pathAPI :: PathParser API