Databrary.Action.Run

Description

How to run Databrary Actions.

This module is the foundation of the site. It provides a method for packaging up route actions into Applications (which is effectively the HTTP transport layer), as well as a method for packaging them into background threads.

"Framework" code

actionApp

Arguments

:: Service
All the low-level system capabilities

-> Action
Action to run

-> Application
Callback for Wai

Convert an Action into a Wai.Application.

This is the 2nd-lowest level of the Databrary "web framework". It runs the requested action with some resolved identity to build the HTTP response. See actionRouteApp for level 1.

forkAction

Arguments

:: Handler a
Handler action to run

-> RequestContext
Original context

The background action inherits this context's Request and Service.

-> (Either SomeException a -> IO ())
Cleanup to run when the action exits

-> IO ThreadId

Run a Handler action in the background (IO).

A new ContextM is built, with a fresh guaranteed db connection and timestamp, via runContextM

Declaring authentication needs for Actions

withAuth

Arguments

:: Handler Response
The handler for the route

-> Action
The bundled action

Tag an Action as needing authentication.

For reference, note that Yesod's equivalent to this function returns Handler Identity, and throws a permission error when the user isn't identified.

withoutAuth :: Handler Response -> Action

Tag an Action as not needing auth. See note at withAuth.

withReAuth

Arguments

:: SiteAuth
The identity to assume

-> Handler a
The action to perform as the assumed identity

Note that one might argue for an indexed form of Handler here, a la Selda and subqueries.

-> Handler a
The re-authenticated action, packaged into the original context

This may be like a su that allows running an action as a different identity.

The underlying type

data Action

This type captures both the authentication needs and the handler for a route.

This type is only found buried within an ActionRoute.